Create failing specs for course group views
